Update a model specification

Update a model specification

# S3 method for discrim_flexible
update(
  object,
  num_terms = NULL,
  prod_degree = NULL,
  prune_method = NULL,
  fresh = FALSE,
  ...
)

# S3 method for discrim_linear
update(
  object,
  penalty = NULL,
  regularization_method = NULL,
  fresh = FALSE,
  ...
)

# S3 method for discrim_quad
update(object, regularization_method = NULL, fresh = FALSE, ...)

# S3 method for discrim_regularized
update(
  object,
  frac_common_cov = NULL,
  frac_identity = NULL,
  fresh = FALSE,
  ...
)

# S3 method for naive_Bayes
update(object, smoothness = NULL, Laplace = NULL, fresh = FALSE, ...)

Arguments

object
A model specification.

num_terms
The number of features that will be retained in the final model, including the intercept.

prod_degree
The highest possible interaction degree.

prune_method
The pruning method.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

...
Not used for update().

penalty
An non-negative number representing the amount of regularization used by some of the engines.

regularization_method
A character string for the type of regularized estimation. Possible values are: "diagonal", "min_distance", "shrink_cov", and "shrink_mean" (sparsediscrim engine only).

frac_common_cov
Numeric values between zero and one.

frac_identity
Numeric values between zero and one.

smoothness
An non-negative number representing the the relative smoothness of the class boundary. Smaller examples result in model flexible boundaries and larger values generate class boundaries that are less adaptable

Laplace
A non-negative value for the Laplace correction to smoothing low-frequency counts.